+++ CHANGES.rst
@@ -5,6 +5,8 @@ Changes in versin 1.1.1
 -----------------------
 
 * Fix accessing fields using ``safe_translation_getter(any_language=True)``
+* Fix "dictionary changed size during iteration" in ``save_translations()`` in Python 3.
+
 
 Changes in version 1.1 (2014-09-29)
 -----------------------------------